Model interface

To simulate your own model, all you have to do is to implement the AbstractModel interface.

StochasticSeriesExpansion.AbstractModel — Type

The type used to define Models that can be simulated by StochasticSeriesExpansion. See StochasticSeriesExpansion.Models.Magnet for a reference implementation.

StochasticSeriesExpansion.generate_sse_data — Function

generate_sse_data(model::AbstractModel) -> SSEData

This function is responsible for generating an SSEData instance for model. It should use the data saved in the model to construct the information needed by the SSE algorithm, such as the bonds, their connection, and the Hamiltonians acting on each bond.

StochasticSeriesExpansion.normalization_site_count — Function

normalization_site_count(model::AbstractModel) -> Int

The number of physical sites count used for normalizing observables. It may differ from the SSE algorithmic site count sometimes, e.g. when the computational basis consists of multiple sites but we still want e.g. our energy to be measured per spin.

StochasticSeriesExpansion.get_opstring_estimators — Function

get_opstring_estimators(model::AbstractModel) -> Vector{Type{<:AbstractOpstringEstimator}}

Returns an array of types of operator string estimators that are used to calculate most observables. Each of them should implement the AbstractOpstringEstimator interface.