OpstringEstimator interface

The AbstractOpstringEstimator interface is used to implement most observable estimators in StochasticSeriesExpansion.

StochasticSeriesExpansion.AbstractOpstringEstimator — Type

AbstractOpstringEstimator

This interface allows defining observable estimators that act on each operator of the the SSE operator string. In the combined measurement the functions of this interface are called like the following.

est = init(YourEstimator, model, state)
for op in operators
    if isidentity(op)
        continue
    end

    if !isdiagonal(op)
        # update state
    end

    if n < num_operators
        measure(est, op, state, sse_data)
    end
    n += 1
end

result(est, mccontext, T, sign)

However, in practice, StochasticSeriesExpansion will interlace different estimators into the same loop for efficiency.

A reference implementation is in the model-generic MagnetizationEstimator, which can be used for most magnetization-like observables.

StochasticSeriesExpansion.init — Function

init(::Type{<:AbstractOpstringEstimator}, model::AbstractModel, state::AbstractVector{<:StateIdx}) -> AbstractOpstringEstimator

Constructs an opstring estimator based on a model and an initial state. The state is a vector of integers labelling the computational site basis states. The estimator needs to interpret them in terms of physical quantities.

StochasticSeriesExpansion.measure — Function

measure(est, op::OperCode, state::AbstractVector{<:StateIdx}, sse_data::SSEData)

Perform the in-string measurement of estimator est on each operator op in the operator string. The state at the current position and the sse_data object are passed for additional reference.

StochasticSeriesExpansion.result — Function

result(est, ctx::Carlo.MCContext, T::AbstractFloat, sign::AbstractFloat)

Finalize the measurement by saving the results to the Carlo MCContext, e.g. by calling measure!(ctx, :Magnetization, est.mag). For more information, consult the Carlo documentation.

For some observables, knowing the temperature T is necessary. In the case of a signful simulation, sign != 1 should be taken into account.

StochasticSeriesExpansion.register_evaluables — Function

register_evaluables(::Type{<:AbstractOpstringEstimator}, eval::Carlo.Evaluator, params::AbstractDict)

Operator string estimators est can define their own evaluables using this function, which passes a Carlo.Evaluator and the task parameters. The state of the estimator is unavailable here since this runs in the postprocessing step.

The reference implementation of this is the general purpose MagnetizationEstimator which works for all models that have some kind of magnetization.

StochasticSeriesExpansion.MagnetizationEstimator — Type

MagnetizationEstimator{OrderingVector, StaggerUC, Model, Prefix, Dimension} <: AbstractOpstringEstimator

Generic operator string estimator that can be used for all models that have

a field lattice::Lattice

implement magnetization_state.

optionally, magnetization_lattice_site_idx.

It computes the

:Mag, magnetization $\langle m\rangle$

:AbsMag, absolute magnetization $\langle |m|\rangle$

:Mag2, squared magnetization $\langle m^2\rangle$

:Mag4, quartic magnetization $\langle m^4\rangle$

:MagChi, susceptibility $N \int_0^\beta d\tau \langle m(\tau) m\rangle$

:BinderRatio, Binder ratio $\langle m^2\rangle^2/\langle m^4\rangle$

Here, $m = \frac{1}{N} \sum_i m_i$ where $m_i$ is given by magnetization_state and $N$ is given by StochasticSeriesExpansion.normalization_site_count.

Type parameters

OrderingVector: Fourier component to compute in units of π. For example $(1,1)$ corresponds to $(π,π)$

If StaggerUC is true, the sublattice sign of the unitcell is additionally taken into account

Model: model type to apply this estimator on

Prefix: (Symbol) this prefix is added to all observable names. Consider magest_standard_prefix.

StochasticSeriesExpansion.magnetization_state — Function

magnetization_state(model::AbstractModel, site_idx, state_idx)

This interface needs to be implemented by any model that wants to use MagnetizationEstimator.

StochasticSeriesExpansion.magnetization_lattice_site_idx — Function

magnetization_lattice_site_idx(model::AbstractModel, sse_site_idx) -> Integer

In models where additional degrees of freedom exist, this function maps sse site indices to physical lattice site indices.

StochasticSeriesExpansion.magest_standard_prefix — Function

magest_standard_prefix(q::Tuple{Bool...}, stagger_uc::Bool)

Returns the conventional prefix for a magnetization estimator with ordering vector q and stagger_uc set.