Simulation data structure

In order to be generic, StochasticSeriesExpansion.jl internally uses data structures that do not know a lot about physics but retain only the necessary information to run the SSE algorithm. As a practicioner, you need to know about these structures either.

If you want to implement your own models or measurements, however, it is your job to translate physics into these data structures, an SSEData object to be precise.

StochasticSeriesExpansion.SSEData — Type

SSEData(vertex_data::AbstractVector{<:VertexData}, bonds::AbstractVector{<:SSEBond})

This object holds everything StochasticSeriesExpansion needs to know to simulate a model using the abstract loop algorithm.

The array vertex_data contains one instance of VertexData for each distinct type of bond. The bonds define the graph of bonds along with the corresponding bond types.

Public fields

bonds: the bond information passed on construction

StochasticSeriesExpansion.get_vertex_data — Function

get_vertex_data(data::SSEData, bond_idx) -> VertexData

The VertexData object for a given SSE bond index bond_idx.

To construct an SSEData object you need to assemble the VertexData and SSEBond objects.

StochasticSeriesExpansion.SSEBond — Type

SSEBond(type, (i, j, ...))

Defines a bond for the SSE algorithm. The type specifies which VertexData object to use. i, j are two or more site indices that are connected by a bond. This number is variable but has to be consistent across the model. If the model has bonds with different number of sites, dummy sites can be used to express lower order bonds.

StochasticSeriesExpansion.VertexData — Type

VertexData(dims::Tuple{Integer...}, bond_hamiltonian::AbstractMatrix; 
           energy_offset_factor = 0.25, tolerance = 1e-7, lp_tolerance = 1e-10)

This object holds the probability tables used for a bond in the abstract loop update algorithm.

dims are the local Hilbert space dimensions of each site of the bond.

bond_hamiltonian is the prod(dims)×prod(dims) matrix describing the bond Hamiltonian in the computational basis. The representation of the product Hilbert space follows the convention of LinearAlgebra.kron, so that if you have site-local operators op1, op2, kron(op1, op2) will give you a correct bond Hamiltonian.

A constant is added to the bond hamiltonian so that its smallest diagonal element is

\[\min_i(h_{ii}) = \alpha \times [\max_i (h_{ii}) - \min_i (h_{ii})] \ge 0,\]

where $\alpha$ is given by the parameter energy_offset_factor.

Matrix elements smaller tolerance are considered zero. lp_tolerance sets the tolerance for the underlying linear programming problem.

StochasticSeriesExpansion.get_leg_state — Function

get_leg_state(vd::VertexData, v::VertexCode)

Returns the leg states (matrix elements) of the vertex v as an array of state indices.